BE/api/pyproject.toml
[project]
name = "gesture-api"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "pydantic-settings>=2.7.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.14.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.3.0",
  "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gesture_api"]
